+++ src/main/resources/static/openapi/campus.json
@@ -460,6 +460,13 @@
     }
   },
   "components": {
+    "securitySchemes": {
+      "Authorization": {
+        "type": "http",
+        "scheme": "bearer",
+        "bearerFormat": "JWT"
+      }
+    },
     "schemas": {
       "MenuItem": {
         "type": "object",
