+++ src/main/resources/static/openapi/info.json
@@ -190,6 +190,13 @@
     }
   },
   "components": {
+    "securitySchemes": {
+      "Authorization": {
+        "type": "http",
+        "scheme": "bearer",
+        "bearerFormat": "JWT"
+      }
+    },
     "schemas": {
       "SearchResult": {
         "type": "object",
