+++ src/main/resources/static/openapi/infra.json
@@ -60,6 +60,13 @@
     }
   },
   "components": {
+    "securitySchemes": {
+      "Authorization": {
+        "type": "http",
+        "scheme": "bearer",
+        "bearerFormat": "JWT"
+      }
+    },
     "schemas": {
       "WeatherResponse": {
         "type": "object",
